Refactor: Move AppLayout and LayoutProvider to layouts directory; reorganize cache components
